pyproject.toml
[project]
name = "pegasus"
version = "0.1.0"
description = "A powerful evaluation framework for Large Language Models using RAGAS and DeepEval"
authors = []
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ragas>=0.0.22",
    "deepeval>=0.11.3",
    "langchain-google-vertexai",
    "pandas>=2.0.0",
    "google-cloud-aiplatform>=1.38.1",
    "google-cloud-storage>=2.13.0",
    "datasets>=2.15.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
select = ["E", "F", "B", "I"]
ignore = ["E501"]
line-length = 88
target-version = "py38"

[tool.hatch.build.targets.wheel]
packages = [".", "evaluator"]
